requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
djangorestframework==3.14.0
ecdsa==0.18.0
pyasn1==0.5.0
pycparser==2.21
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1

psycopg2-binary
gunicorn